Allow field type without inheritance of extends.
This allows us to buypass some of the default field setups and allows to render single fields.  Really the current field structure is for fields that exist in the cardwall in a prebuilt layout structure.  We might not want that outer markup so this partial implementation allows us to avoid some of the extra markup and jsut render a single input.