Fix buttons alignment for pickers
Note this hardcodes the input-group append and prepend styling for beta, and priovides the styling.  Prepend and append are not actually supported for these fields at this time due to some custom data that is set in the data-g5-filepicker field. When trying to generalize the output (controlled by YAML), it was found we may need to set the data-g5-filepicker differenlty.