Update logo.html.twig
May be useful to have a hard coded class in some particles. Lets say from aesthetic reasons you want to limit the logo image size in navigation section. Of course you can add a class in particle settings, but this way, that class inadvertently can be removed. 

This is template specific, but i think will be just fine to have that class in there. You have the last word :D .