Removed base theme import from dependencies (#1152)
The "nucleus/theme/base" style has never been a set of mixins / functions, therefore when loading `dependencies` through custom.scss, plain CSS would get added in the compiled css, adding a lot of unnecessary duplicated code.

This change takes away the theme base from dependencies and moves it at the top of the theme.scss itself.

## Themes Developers
If you are a theme developer and you have this setup, make sure you update your themes as well. This is supposedly still working fine if you don't syncrhonize but it is a bug.