Fixing minor Modal bug. (#1454)
Different browsers with padding calculate heights different. Firefox and Chrome have an inconsistency of display here. Chrome respects the 10vh padding-bottom for the g5-dialog box. Firefox calculates heights differently, as such the 10vh is never adding to the bottom of the modal. To force this to render properly, because of how the box-model works with margin properties as being “addition” to heights, we convert the code to margin use. Enforcing this as an “addition” height guarentees the modal will render with the viewport margin (in additon to the modal height based on children).