Added aria-label to the Social Particle
I added `aria-label="{{ item.text|e }}"` inside <a> in the
social.html.twig file so it looks like this.
```twig
<a target="{{ particle.target|default('_blank')|e }}" href="{{
item.link|e }}" title="{{ item.text|e }}" aria-label="{{ item.text|e
}}">
{% if particle.display in ['both', 'icons_only'] %}<span class="{{
item.icon|e }}"></span>{% endif %}
{% if particle.display in ['both', 'text_only'] %}<span
class="g-social-text">{{ item.text|e }}</span>{% endif %}
</a>
```

By adding it this way, if text is added to the **Title** field, it will
also be picked up and added to `aria-label`. This prevents the necessity
of adding a separate field for aria-label and takes care of adding it
automatically but only if text is entered in the **Title** field.

A use case scenario where this is necessary, from an accessibility stand
point, is that if **icons only** is selected, which means *no text for
the link*, screen readers will see it as an empty link. This addition
takes care of that problem.